requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
email-validator==2.0.0.post2
requests==2.31.0
Jinja2==3.1.2
itsdangerous==2.1.2
cryptography==41.0.2
